:recycle: Add exception management - Add specific exceptions - fix transactionals - add validations on Dtos

+++ backend/src/main/java/com/example/eventmanager/application/services/EventServiceImpl.java
@@ -8,6 +8,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.example.eventmanager.domain.exceptions.EventNotFoundException;
+import com.example.eventmanager.domain.exceptions.ProductNotFoundException;
 import com.example.eventmanager.domain.models.Event;
 import com.example.eventmanager.domain.models.Product;
 import com.example.eventmanager.domain.repositories.EventRepository;
@@ -30,9 +32,10 @@ public class EventServiceImpl implements EventService {
     }
 
     @Override
+    @Transactional
     public Event createEvent(EventCreationRequest eventCreationRequest) {
         Product product = productRepository.findDomainProductById(eventCreationRequest.getProductId())
-            .orElseThrow(() -> new RuntimeException("Product not found"));
+            .orElseThrow(() -> new ProductNotFoundException(eventCreationRequest.getProductId()));
         
         LocalDate startDate = (eventCreationRequest.getDate() != null) ? eventCreationRequest.getDate() : LocalDate.now();
 
@@ -52,16 +55,26 @@ public class EventServiceImpl implements EventService {
     }
 
     @Override
+    @Transactional(readOnly = true)
     public List<Event> findDomainEventByProductId(UUID productId) {
-        return eventRepository.findDomainEventByProductId(productId);
+        List<Event> events = eventRepository.findDomainEventByProductId(productId);
+        if (events.isEmpty()) {
+            throw new EventNotFoundException(productId, "No events found for this product");
+        }
+        return events;
     }
 
     @Override
+    @Transactional
     public void deleteById(UUID id) {
+        if (!eventRepository.existsById(id)) {
+            throw new EventNotFoundException(id);
+        }
         eventRepository.deleteById(id);
     }
 
     @Override
+    @Transactional(readOnly = true)
     public boolean existsById(UUID id) {
         return eventRepository.existsById(id);
     }
